Added user input
I added the try clause and to have user input and be able to solve issues with making the input only accept integers. I also used global variables as it allows me to change the variable inside of a function and enclose the try clause inside of a function. The reason why I added these changes was to have user input and make them be able to change the shape. The shape right now is a bunch of lines and I will change that soon.