Changed Shape
Changed the shape to draw hexagons. I also added more to the color list in choices and I found a way to increase the speed of turtle beyond the turtle.speed(). It is pretty fast and if you want to see the turtle draw the shapes, you can comment those out.  There is also another shape option and you can remove the comment on line 18 to activate that.